refactor: 파일 및 컴포넌트 이름 수정
- make-quiz-dialog => make-quiz-dialog-content
- make-quiz-drawer => make-quiz-drawer-content
- make-quiz-drawer-contents => make-quiz-drawer-steps
